datactl: Make FloatSignalBlock type 32-bit
This is more in-line with what DAQ systems actually produce and has some
algorithmic advantages.

It also gets both IntSignalBlock and FloatSignalBlock datatypes nicely
aligned to be 32-bit precise, which should lead to less surprises when
we add data types with other precisions later.

+++ modules/plot-timeseries/timeplotwidget.cpp
@@ -199,7 +199,7 @@ int TimePlotWidget::addSeries(const QString &seriesName, const PlotSeriesSetting
     return d->xdata.size() - 1;
 }
 
-void TimePlotWidget::addToSeriesF(int seriesIndex, const Eigen::RowVectorXd &vec)
+void TimePlotWidget::addToSeriesF(int seriesIndex, const Eigen::RowVectorXf &vec)
 {
     const std::lock_guard<std::mutex> lock(d->dataMutex);
